PrioritySetActivity it's working including update contactMap!! change contact id from int to long

+++ keepintouch/app/src/main/java/com/example/keepintouch/MainActivity.java
@@ -122,7 +122,7 @@ public class MainActivity extends AppCompatActivity {
                 tvName.setText(contact.getName());
                 tvNumber.setText(contact.getNumber());
                 ivView.setImageURI(Uri.parse(contact.getPhotoSrc()));
-                tvId.setText(contact.getContactId());
+                tvId.setText((int) contact.getContactId());
                 return convertView;
             }
         };
